# openapi/management-open-api.yaml
openapi: 3.0.3
info:
  title: iceberg-catalog
  description: |
    Implementation of the Iceberg REST Catalog server.
  license:
    name: Apache-2.0
  version: 0.4.0
paths:
...
  /management/v1/warehouse:
    get:
      tags:
      - management
      summary: List all warehouses in a project
      description: |-
        By default, this endpoint does not return deactivated warehouses.
        To include deactivated warehouses, set the `include_deactivated` query parameter to `true`.
      operationId: list_warehouses
      parameters:
      - name: warehouse-status
        in: query
        description: |-
          Optional filter to return only warehouses
          with the specified status.
          If not provided, only active warehouses are returned.
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/WarehouseStatus'
          nullable: true
      - name: project-id
        in: query
        description: |-
          The project ID to list warehouses for.
          Setting a warehouse is required.
        required: false
        schema:
          type: string
          format: uuid
          nullable: true
      responses:
        '200':
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListWarehousesResponse'
    post:
      tags:
      - management
      summary: Create a new warehouse.
      description: |-
        Create a new warehouse in the given project. The project
        of a warehouse cannot be changed after creation.
        The storage configuration is validated by this method.
      operationId: create_warehouse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWarehouseRequest'
        required: true
      responses:
        '201':
          description: Warehouse created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreateWarehouseResponse'
  /management/v1/warehouse/{warehouse_id}:
    get:
      tags:
      - management
      summary: Get a warehouse by ID
      operationId: get_warehouse
      parameters:
      - name: warehouse_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Warehouse details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetWarehouseResponse'
    delete:
      tags:
      - management
      summary: Delete a warehouse by ID
      operationId: delete_warehouse
      parameters:
      - name: warehouse_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Warehouse deleted successfully
  /management/v1/warehouse/{warehouse_id}/activate:
    post:
      tags:
      - management
      summary: Activate a warehouse
      operationId: activate_warehouse
      parameters:
      - name: warehouse_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Warehouse activated successfully
  /management/v1/warehouse/{warehouse_id}/deactivate:
    post:
      tags:
      - management
      summary: Deactivate a warehouse
      operationId: deactivate_warehouse
      parameters:
      - name: warehouse_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Warehouse deactivated successfully
...
  /management/v1/warehouse/{warehouse_id}/rename:
    post:
      tags:
      - management
      summary: Rename a warehouse
      operationId: rename_warehouse
      parameters:
      - name: warehouse_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameWarehouseRequest'
        required: true
      responses:
        '200':
          description: Warehouse renamed successfully
  /management/v1/warehouse/{warehouse_id}/storage:
    post:
      tags:
      - management
      summary: Update the storage profile of a warehouse including its storage credential.
      operationId: update_storage_profile
      parameters:
      - name: warehouse_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWarehouseStorageRequest'
        required: true
      responses:
        '200':
          description: Storage profile updated successfully
  /management/v1/warehouse/{warehouse_id}/storage-credential:
    post:
      tags:
      - management
      summary: Update the storage credential of a warehouse. The storage profile is not modified.
      description: This can be used to update credentials before expiration.
      operationId: update_storage_credential
      parameters:
      - name: warehouse_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWarehouseCredentialRequest'
        required: true
      responses:
        '200':
          description: Storage credential updated successfully
components:
  schemas:
    AzCredential:
      oneOf:
      - type: object
        required:
        - client-id
        - tenant-id
        - client-secret
        - credential-type
        properties:
          client-id:
            type: string
          client-secret:
            type: string
          credential-type:
            type: string
            enum:
            - client-credentials
          tenant-id:
            type: string
      discriminator:
        propertyName: credential-type
...
    CreateWarehouseRequest:
      type: object
      required:
      - warehouse-name
      - storage-profile
      properties:
        delete-profile:
          $ref: '#/components/schemas/TabularDeleteProfile'
        project-id:
          type: string
          format: uuid
          description: |-
            Project ID in which to create the warehouse.
            If no default project is set for this server, this field is required.
          nullable: true
        storage-credential:
          allOf:
          - $ref: '#/components/schemas/StorageCredential'
          nullable: true
        storage-profile:
          $ref: '#/components/schemas/StorageProfile'
        warehouse-name:
          type: string
          description: |-
            Name of the warehouse to create. Must be unique
            within a project and may not contain "/"
    CreateWarehouseResponse:
      type: object
      required:
      - warehouse-id
      properties:
        warehouse-id:
          type: string
          format: uuid
          description: ID of the created warehouse.
...
    GcsProfile:
      type: object
      required:
      - bucket
      properties:
        bucket:
          type: string
          description: Name of the GCS bucket
        key-prefix:
          type: string
          description: |-
            Subpath in the bucket to use.
            The same prefix can be used for multiple warehouses.
          nullable: true
...
    GetWarehouseResponse:
      type: object
      required:
      - id
      - name
      - project-id
      - storage-profile
      - status
      properties:
        id:
          type: string
          format: uuid
          description: ID of the warehouse.
        name:
          type: string
          description: Name of the warehouse.
        project-id:
          type: string
          format: uuid
          description: Project ID in which the warehouse is created.
        status:
          $ref: '#/components/schemas/WarehouseStatus'
        storage-profile:
          $ref: '#/components/schemas/StorageProfile'
...
    ListWarehousesResponse:
      type: object
      required:
      - warehouses
      properties:
        warehouses:
          type: array
          items:
            $ref: '#/components/schemas/GetWarehouseResponse'
          description: List of warehouses in the project.
    RenameWarehouseRequest:
      type: object
      required:
      - new-name
      properties:
        new-name:
          type: string
          description: New name for the warehouse.
    S3Credential:
      oneOf:
      - type: object
        required:
        - aws-access-key-id
        - aws-secret-access-key
        - credential-type
        properties:
          aws-access-key-id:
            type: string
          aws-secret-access-key:
            type: string
          credential-type:
            type: string
            enum:
            - access-key
      discriminator:
        propertyName: credential-type
...
    StorageCredential:
      oneOf:
      - allOf:
        - $ref: '#/components/schemas/S3Credential'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - s3
      - allOf:
        - $ref: '#/components/schemas/AzCredential'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - az
      - allOf:
        - $ref: '#/components/schemas/GcsCredential'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - gcs
      description: Storage secret for a warehouse.
      discriminator:
        propertyName: type
    StorageProfile:
      oneOf:
      - allOf:
        - $ref: '#/components/schemas/AzdlsProfile'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - azdls
      - allOf:
        - $ref: '#/components/schemas/S3Profile'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - s3
      - allOf:
        - $ref: '#/components/schemas/GcsProfile'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - gcs
      description: Storage profile for a warehouse.
      discriminator:
        propertyName: type
    TabularDeleteProfile:
      oneOf:
      - type: object
        required:
        - type
        properties:
          type:
            type: string
            enum:
            - hard
      - type: object
        required:
        - expiration_seconds
        - type
        properties:
          expiration_seconds:
            type: string
          type:
            type: string
            enum:
            - soft
      discriminator:
        propertyName: type
...
    UpdateWarehouseCredentialRequest:
      type: object
      properties:
        new-storage-credential:
          allOf:
          - $ref: '#/components/schemas/StorageCredential'
          nullable: true
    UpdateWarehouseStorageRequest:
      type: object
      required:
      - storage-profile
      properties:
        storage-credential:
          allOf:
          - $ref: '#/components/schemas/StorageCredential'
          nullable: true
        storage-profile:
          $ref: '#/components/schemas/StorageProfile'
    WarehouseStatus:
      type: string
      description: Status of a warehouse
      enum:
      - active
      - inactive
...